pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wisperflow"
version = "0.1.0"
description = "Voice-to-text macOS app powered by OpenAI Whisper. Hold or toggle a shortcut to record, transcribe, and auto-paste."
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "openai-whisper",
    "torch",
    "numpy",
    "sounddevice",
    "scipy",
    "pynput",
    "pywebview",
    "pystray",
    "Pillow",
    "pyperclip",
]

[project.scripts]
wisperflow = "wisperflow.__main__:main"

[tool.setuptools.packages.find]
include = ["wisperflow*"]

[tool.setuptools.package-data]
wisperflow = ["assets/*"]
